componentes de condusta de datos organizado por componentes separados a falta de mostrar en el dashboard
