+++ src/components/FetchData.jsx
@@ -1,0 +1,96 @@
+import dayjs from 'dayjs';
+import { CalendarOnly } from './calendar';
+import { useState } from 'react';
+import { apiUser } from '../api/client';
+import { Button, Stack, Grid, Typography} from '@mui/material';
+
+import { Entries } from './Entries';
+import { Sums } from './Sums';
+import { Average } from './Averages';
+import { TotalsCards } from './TotalsCards';
+
+
+
+export const EntriesList = ({ }) => {
+
+
+    const [startDate, setStartDate] = useState(dayjs());
+    const [endDate, setEndDate] = useState(dayjs());
+
+
+    const [entries, setEntries] = useState([])
+    const [loading, setLoading] = useState(false)
+    const [error, setError] = useState('')
+    const [totals, setTotals] = useState({ total_entries: 0, averages_by_shift: {}, sums_by_shift: {} });
+
+    const fetchEntries = async () => {
+
+        setLoading(true);
+        setError('');
+        try {
+
+            const data = await apiUser(`/entries/range/?start_date=${startDate.format('YYYY-MM-DD')}&end_date=${endDate.format('YYYY-MM-DD')}`)
+            setEntries(data.entries || [])
+            setTotals({
+                total_entries: data.total_entries || 0,
+                averages_by_shift: data.averages_by_shift || {},
+                sums_by_shift: data.sums_by_shift || {}
+            });
+
+        } catch (err) {
+            console.error(err);
+            setError('Error al cargar las entradas');
+        } finally {
+            setLoading(false);
+        }
+    };
+
+    return (
+
+        <Stack spacing={2} sx={{ p: 2 }}>
+            {/* Fechas y botón */}
+            <Grid container spacing={2} alignItems="center">
+                <Grid item xs={12} sm={5}>
+                    <CalendarOnly calendarDay={(day) => setStartDate(day)} />
+                </Grid>
+                <Grid item xs={12} sm={5}>
+                    <CalendarOnly calendarDay={(day) => setEndDate(day)} />
+                </Grid>
+                <Grid item xs={12} sm={2}>
+                    <Button fullWidth variant="contained" onClick={fetchEntries}>
+                        Buscar entradas
+                    </Button>
+                </Grid>
+            </Grid>
+
+            {/* Totales */}
+            <Grid container spacing={2}>
+                <Grid item xs={12} sm={4}>
+                    <TotalsCards total={totals.total_entries} />
+                </Grid>
+            </Grid>
+
+            {/* Promedios */}
+            <Grid container spacing={2}>
+                <Average average={totals.averages_by_shift} />
+            </Grid>
+
+            {/* Sumatorios */}
+            <Grid container spacing={2}>
+                <Sums sums={totals.sums_by_shift} />
+            </Grid>
+
+            {/* Entradas individuales */}
+            <Entries entries={entries} />
+
+
+            {loading && <Typography>Cargando...</Typography>}
+            {error && <Typography color="error">{error}</Typography>}
+        </Stack>
+    );
+}
+
+
+
+
+
